+++ resources/views/components/icon.blade.php
@@ -75,6 +75,9 @@
         @case('chevron-down')
             <path d="m6 9 6 6 6-6"/>
             @break
+        @case('chevron-right')
+            <path d="m9 18 6-6-6-6"/>
+            @break
         @case('rotate-ccw')
             <path d="M3 12a9 9 0 1 0 9-9 9.75 9.75 0 0 0-6.74 2.74L3 8"/><path d="M3 3v5h5"/>
             @break